Add basic typing to emulated_hue
