+++ homeassistant/components/emulated_hue/const.py
@@ -2,3 +2,5 @@
 
 HUE_SERIAL_NUMBER = "001788FFFE23BFC2"
 HUE_UUID = "2f402f80-da50-11e1-9b23-001788255acc"
+
+DOMAIN = "emulated_hue"
