+++ tests/components/emulated_hue/test_upnp.py
@@ -2,6 +2,7 @@
 from http import HTTPStatus
 import json
 import unittest
+from unittest.mock import patch
 
 from aiohttp import web
 import defusedxml.ElementTree as ET
@@ -52,11 +53,13 @@ def hue_client(aiohttp_client):
 
 async def setup_hue(hass):
     """Set up the emulated_hue integration."""
-    assert await setup.async_setup_component(
-        hass,
-        emulated_hue.DOMAIN,
-        {emulated_hue.DOMAIN: {emulated_hue.CONF_LISTEN_PORT: BRIDGE_SERVER_PORT}},
-    )
+    with patch("homeassistant.components.emulated_hue.create_upnp_datagram_endpoint"):
+        assert await setup.async_setup_component(
+            hass,
+            emulated_hue.DOMAIN,
+            {emulated_hue.DOMAIN: {emulated_hue.CONF_LISTEN_PORT: BRIDGE_SERVER_PORT}},
+        )
+        await hass.async_block_till_done()
 
 
 def test_upnp_discovery_basic():
